+++ source-code/cljs/app_plugins/item_browser/api.cljs
@@ -28,8 +28,7 @@
 ;; ----------------------------------------------------------------------------
 
 ; app-plugins.item-browser.engine
-(def browser-uri engine/browser-uri)
-(def request-id  engine/request-id)
+(def request-id engine/request-id)
 
 ; app-plugins.item-browser.events
 (def toggle-item-selection! events/toggle-item-selection!)