+++ source-code/cljs/app_plugins/item_browser/engine.cljs
@@ -12,19 +12,12 @@
 ;; ----------------------------------------------------------------------------
 
 ; mid-plugins.item-browser.engine
-(def browser-uri             engine/browser-uri)
-(def request-id              engine/request-id)
-(def mutation-name           engine/mutation-name)
-(def resolver-id             engine/resolver-id)
-(def collection-name         engine/collection-name)
-(def transfer-id             engine/transfer-id)
-(def route-id                engine/route-id)
-(def extended-route-id       engine/extended-route-id)
-(def route-template          engine/route-template)
-(def extended-route-template engine/extended-route-template)
-(def component-id            engine/component-id)
-(def dialog-id               engine/dialog-id)
-(def load-extension-event    engine/load-extension-event)
+(def request-id      engine/request-id)
+(def mutation-name   engine/mutation-name)
+(def resolver-id     engine/resolver-id)
+(def collection-name engine/collection-name)
+(def component-id    engine/component-id)
+(def dialog-id       engine/dialog-id)