+++ source-code/cljs/app_plugins/item_browser/subs.cljs
@@ -8,7 +8,6 @@
               [mid-fruits.candy                :refer [param return]]
               [mid-fruits.keyword              :as keyword]
               [mid-fruits.loop                 :refer [some-indexed]]
-              [mid-plugins.item-browser.subs   :as subs]
               [x.app-core.api                  :as a :refer [r]]
               [x.app-db.api                    :as db]
               [x.app-router.api                :as router]))
@@ -24,14 +23,31 @@
 (def lister-disabled?       app-plugins.item-lister.subs/lister-disabled?)
 (def toggle-item-selection? app-plugins.item-lister.subs/toggle-item-selection?)
 
-; mid-plugins.item-browser.subs
-(def get-browser-props subs/get-browser-props)
-(def get-meta-item     subs/get-meta-item)
-
 
 
 ;; ----------------------------------------------------------------------------
 ;; ----------------------------------------------------------------------------
+
+(defn get-browser-props
+  ; WARNING! NON-PUBLIC! DO NOT USE!
+  ;
+  ; @param (keyword) extension-id
+  ; @param (keyword) item-namespace
+  ;
+  ; @return (map)
+  [db [_ extension-id _]]
+  (get-in db [extension-id :item-browser/meta-items]))
+
+(defn get-meta-item
+  ; WARNING! NON-PUBLIC! DO NOT USE!
+  ;
+  ; @param (keyword) extension-id
+  ; @param (keyword) item-namespace
+  ; @param (keyword) item-key
+  ;
+  ; @return (*)
+  [db [_ extension-id item-namespace item-key]]
+  (get-in db [extension-id :item-browser/meta-items item-key]))
 
 (defn get-current-item-id
   ; @param (keyword) extension-id
@@ -177,10 +193,10 @@
   ; @param (keyword) item-namespace
   ;
   ; @return (boolean)
-  [db [_ extension-id item-namespace]]
-  (let [route-id (r router/get-current-route-id db)]
-       (or (= route-id (engine/route-id          extension-id item-namespace))
-           (= route-id (engine/extended-route-id extension-id item-namespace)))))
+  [db [_ extension-id item-namespace]])
+  ;(let [route-id (r router/get-current-route-id db)]
+  ;     (or (= route-id (engine/route-id          extension-id item-namespace))
+  ;         (= route-id (engine/extended-route-id extension-id item-namespace))]])
 
 (defn set-title?
   ; WARNING! NON-PUBLIC! DO NOT USE!
@@ -224,6 +240,14 @@
 
 ;; ----------------------------------------------------------------------------
 ;; ----------------------------------------------------------------------------
+
+; @param (keyword) extension-id
+; @param (keyword) item-namespace
+; @param (keyword) item-key
+;
+; @usage
+;  [:item-browser/get-meta-item :my-extension :my-type :my-item]
+(a/reg-sub :item-browser/get-meta-item get-meta-item)
 
 ; @param (keyword) extension-id
 ; @param (keyword) item-namespace